Optimize Database Schema

+++ src/main/java/com/anshuman/workflow/data/dto/WorkflowEventLogDto.java
@@ -16,7 +16,7 @@ public class WorkflowEventLogDTO {
 
     private Long companyId;
 
-    private Long branchId;
+    private Integer branchId;
 
     private WorkflowType typeId;
